get_db_age_nodes can now return clade names and the Euk+Bac category.
Previously, I was writing age dictonaries that had only the MRCA node for each
protein family, but this threw away the Euk+Bac category, which is not
monophyletic. This new functionality writes the age categories at the
serialization step instead of after, so it will replace the nodes2taxonAge
function in analysis.py

The solution was pretty hacky, involving a hard-coded list of Archaea, but
I'm okay with it for now.